+++ client/coral-admin/src/components/Indicator.css
