+++ client/coral-admin/src/components/Indicator.js
