+++ client/coral-admin/src/components/Layout.css
