+++ client/coral-admin/src/components/Logo.css
