+++ client/coral-admin/src/components/Logo.js
